Flyttet periodeliste inn i egen komponent
